fix: handle an issue when using urls as part of a target value
When using a URL (or anything else with a slash) as part of a target
value, then this may lead to a corrupt URL.

The reason for this is that it will actually create a different
hierarchy, one that is no longer understood by the parser.

The fix to this problem is to "URL encode" the values, as they become
segments of the path.

BREAKING-CHANGE: Target values are not URL encoded. This might change
the path which is being generated. The previous behavior wasn't reliable
however.