feat: allow pushing the state alongside the target
BREAKING-CHANGE: Instead of using gloo_history, this now uses the plain
browser APIs. This also pushes the actual value to the page state.
Instead of relying on the internal state array. This means that while
you can get the value as actual JsValue now, it is not possible to
use it in combination with gloo_history.